requirements.txt
tensorflow
tensorflow-hub
numpy
scipy
google-generativeai
python-dotenv

